examples: moving config.json to shared location
