+++ examples/graph/index.ts
@@ -13,7 +13,7 @@ import { GossipManager } from "@lntools/wire";
 import fs from "fs";
 
 // tslint:disable-next-line: no-var-requires
-const config = require("./config.json");
+const config = require("../config.json");
 
 const logger = new Logger("app");
 logger.transports.push(new ConsoleTransport(console));
